+++ cedar-policy-symcc/CHANGELOG.md
@@ -11,7 +11,6 @@ Cedar Language Version: TBD
 ### Added
 
 - `Display` impl for `Env`
-- Fixed bug in `Term::interpret` caused by missing `bvsrem` case
 
 ## [0.3.0] - 2026-02-17
 Cedar Language Version: 4.4
@@ -44,6 +43,7 @@ use the performance-optimized primitives introduced in 0.2.0. (#2102)
 
 - Bug where returned counterexamples could occasionally be invalid by containing
 cycles in the entity data for entities irrelevant to the given policies (#2089)
+- Fixed bug in `Term::interpret` caused by missing `bvsrem` case
 
 ### Removed